Added SkyMaterial to remove the getEmittance method from the Material
This change opened the way for some more improvements: the masked
material doesn't need to rely on the pass number anymore for consistency
because there is only one method to call now. This also means the
Context can be simplified. Scenes that don't need a sky can now just
ignore it because the Scene class defaults to a black sky.